extract the ported examples to their own test module
