+++ src/fact_emitter/examples/issue_47680.rs
