+++ src/fact_emitter/examples/vec_temp.rs
